Rename parseMembers -> parseUsers
