+++ apps/bot/src/util/parseUsers.ts
